Quick adjust for Mobile and Web

+++ src/main/java/com/example/thanhhoa/services/vnpay/VNPayService.java
@@ -4,7 +4,7 @@ import javax.servlet.http.HttpServletRequest;
 
 public interface VNPayService {
 
-    String createOrder(int total, String reason, String urlReturn, String ipAddr);
+    String createOrder(Double total, String reason, String urlReturn, String ipAddr);
 
     int orderReturn(HttpServletRequest request);
 }
